+++ Framework/DataHandling/src/LoadHFIRSANS.cpp
@@ -21,7 +21,6 @@
 #include "MantidKernel/TimeSeriesProperty.h"
 #include "MantidKernel/UnitFactory.h"
 
-#include <boost/algorithm/string/split.hpp>
 #include <boost/algorithm/string/trim.hpp>
 #include <boost/lexical_cast.hpp>
 #include <boost/make_shared.hpp>
@@ -47,6 +46,10 @@
 #include <string>
 #include <utility>
 #include <vector>
+
+using Poco::XML::Document;
+using Poco::XML::DOMParser;
+using Poco::XML::Element;
 
 namespace Mantid {
 namespace DataHandling {
@@ -78,8 +81,8 @@ int LoadHFIRSANS::confidence(Kernel::FileDescriptor &descriptor) const {
   { // start of inner scope
     Poco::XML::InputSource src(is);
     // Set up the DOM parser and parse xml file
-    Poco::XML::DOMParser pParser;
-    Poco::AutoPtr<Poco::XML::Document> pDoc;
+    DOMParser pParser;
+    Poco::AutoPtr<Document> pDoc;
     try {
       pDoc = pParser.parse(&src);
     } catch (Poco::Exception &e) {
@@ -91,7 +94,7 @@ int LoadHFIRSANS::confidence(Kernel::FileDescriptor &descriptor) const {
                                          descriptor.filename());
     }
     // Get pointer to root element
-    Poco::XML::Element *pRootElem = pDoc->documentElement();
+    Element *pRootElem = pDoc->documentElement();
     if (pRootElem) {
       if (pRootElem->tagName() == "SPICErack") {
         confidence = 80;
@@ -112,8 +115,8 @@ void LoadHFIRSANS::init() {
                   "The name of the Output workspace");
 
   // Optionally, we can specify the wavelength and wavelength spread and
-  // overwrite the value in the data file (used when the data file is not
-  // populated)
+  // overwrite
+  // the value in the data file (used when the data file is not populated)
   auto mustBePositive = boost::make_shared<Kernel::BoundedValidator<double>>();
   mustBePositive->setLower(0.0);
   declareProperty("Wavelength", EMPTY_DBL(), mustBePositive,
@@ -140,16 +143,16 @@ void LoadHFIRSANS::exec() {
   setWavelength();
   createWorkspace();
   storeMetaDataIntoWS();
+  runLoadInstrument();
   // ugly hack for Biosans wing detector:
   // it tests if there is metadata tagged with the wing detector
   // if so, puts the detector in the right angle
   if (m_metadata.find("Motor_Positions/det_west_wing_rot") !=
       m_metadata.end()) {
     rotateDetector();
+    runLoadInstrument();
   }
   moveDetector();
-  runLoadInstrument();
-  // This needs parameters from IDF! Run load instrument before!
   setBeamDiameter();
   setProperty("OutputWorkspace", m_workspace);
 }
@@ -321,6 +324,7 @@ std::vector<int> LoadHFIRSANS::getData(const std::string &dataXpath) {
  * Convenience function to store a detector value into a given spectrum.
  * Note that this type of data doesn't use TOD, so that we use a single dummy
  * bin in X. Each detector is defined as a spectrum of length 1.
+ * @param ws: workspace
  * @param specID: ID of the spectrum to store the value in
  * @param value: value to store [count]
  * @param error: error on the value [count]
@@ -380,20 +384,7 @@ void LoadHFIRSANS::createWorkspace() {
 template <class T>
 void LoadHFIRSANS::addRunProperty(const std::string &name, const T &value,
                                   const std::string &units) {
-  g_log.debug() << "Adding Property to the Run: " << name << " -> " << value
-                << "\n";
   m_workspace->mutableRun().addProperty(name, value, units, true);
-}
-
-template <class T>
-void LoadHFIRSANS::addRunTimeSeriesProperty(const std::string &name,
-                                            const T &value) {
-  g_log.debug() << "Adding Time Series Property to the Run: " << name << " -> "
-                << value << "\n";
-  API::Run &runDetails = m_workspace->mutableRun();
-  auto *p = new Mantid::Kernel::TimeSeriesProperty<T>(name);
-  p->addValue(DateAndTime::getCurrentTime(), value);
-  runDetails.addLogData(p);
 }
 
 /**
@@ -487,7 +478,8 @@ void LoadHFIRSANS::storeMetaDataIntoWS() {
   addRunProperty<double>(
       "timer", boost::lexical_cast<double>(m_metadata["Counters/time"]), "sec");
 
-  // XML 1.03: sample thickness is now in meters
+  // sample thickness
+  // XML 1.03: source distance is now in meters
   double sample_thickness =
       boost::lexical_cast<double>(m_metadata["Header/Sample_Thickness"]);
   if (m_sansSpiceXmlFormatVersion >= 1.03) {
@@ -505,6 +497,15 @@ void LoadHFIRSANS::storeMetaDataIntoWS() {
       "sample-aperture-diameter",
       boost::lexical_cast<double>(m_metadata["Header/sample_aperture_size"]),
       "mm");
+  // XML 1.03: source distance is now in meters
+  double source_distance =
+      boost::lexical_cast<double>(m_metadata["Header/source_distance"]);
+  if (m_sansSpiceXmlFormatVersion >= 1.03) {
+    g_log.debug() << "sans_spice_xml_format_version >= 1.03 :: source_distance "
+                     "in meters. Converting to mm...";
+    source_distance *= 1000.0;
+  }
+  addRunProperty<double>("source-sample-distance", source_distance, "mm");
   addRunProperty<double>(
       "number-of-guides",
       boost::lexical_cast<double>(m_metadata["Motor_Positions/nguides"]));
@@ -538,18 +539,43 @@ void LoadHFIRSANS::runLoadInstrument() {
 
 /**
  * This will rotate the detector named componentName around z-axis
+ *
+ * @param angle in degrees
  */
 void LoadHFIRSANS::rotateDetector() {
 
   // The angle is negative!
   double angle = -boost::lexical_cast<double>(
       m_metadata["Motor_Positions/det_west_wing_rot"]);
+
   g_log.notice() << "Rotating Wing Detector " << angle << " degrees." << '\n';
-  addRunTimeSeriesProperty<double>("rotangle", angle);
+
+  API::Run &runDetails = m_workspace->mutableRun();
+  auto *p = new Mantid::Kernel::TimeSeriesProperty<double>("rotangle");
+  //	auto p = boost::make_shared <Mantid::Kernel::TimeSeriesProperty<double>
+  //>("rotangle");
+  p->addValue(DateAndTime::getCurrentTime(), angle);
+  runDetails.addLogData(p);
 }
 
 /**
  * Calculates the detector distances and sets them as Run properties
+ * Here fog starts:
+ * GPSANS: distance = sample_det_dist + offset!
+ * BioSANS: distance = sample_det_dist + offset + sample_to_flange!
+ * Mathieu is using sample_det_dist to move the detector later
+ * So I'll do the same (Ricardo)
+ * June 14th 2016:
+ * New changes:
+ * sample_det_dist is not available
+ * flange_det_dist is new = old sample_det_dist + offset
+ * offset is not used
+ * GPSANS: distance = flange_det_dist! (sample_to_flange is 0 for GPSANS)
+ * BioSANS: distance = flange_det_dist + sample_to_flange!
+ * For back compatibility I'm setting the offset to 0 and not reading it from
+ * the file
+ * Last Changes:
+ * If SDD tag is available in the metadata set that as sample detector distance
  * @return : sample_detector_distance
  */
 void LoadHFIRSANS::setDetectorDistance() {
@@ -592,7 +618,7 @@ void LoadHFIRSANS::setDetectorDistance() {
                 << " mm." << '\n';
   addRunProperty<double>("sample-detector-distance", m_sampleDetectorDistance,
                          "mm");
-  addRunTimeSeriesProperty<double>("sdd", m_sampleDetectorDistance);
+  // return m_sampleDetectorDistance;
 }
 
 /**
@@ -601,11 +627,36 @@ void LoadHFIRSANS::setDetectorDistance() {
 void LoadHFIRSANS::moveDetector() {
 
   setDetectorDistance();
-  double translationDistance =
+  double translation_distance =
       boost::lexical_cast<double>(m_metadata["Motor_Positions/detector_trans"]);
-  g_log.debug() << "Detector Translation = " << translationDistance << " mm."
-                << '\n';
-  addRunTimeSeriesProperty<double>("detector-translation", translationDistance);
+  translation_distance /= 1000.0;
+  g_log.debug() << "Detector Translation = " << translation_distance
+                << " meters." << '\n';
+
+  // Move the detector to the right position
+  API::IAlgorithm_sptr mover = createChildAlgorithm("MoveInstrumentComponent");
+
+  // Finding the name of the detector object.
+  std::string detID =
+      m_workspace->getInstrument()->getStringParameter("detector-name")[0];
+
+  g_log.information() << "Moving: " << detID
+                      << " Z=" << m_sampleDetectorDistance / 1000.0
+                      << " X=" << translation_distance << '\n';
+  try {
+    mover->setProperty<API::MatrixWorkspace_sptr>("Workspace", m_workspace);
+    mover->setProperty("ComponentName", detID);
+    mover->setProperty("Z", m_sampleDetectorDistance / 1000.0);
+    mover->setProperty("X", -translation_distance);
+    mover->execute();
+  } catch (std::invalid_argument &e) {
+    g_log.error("Invalid argument to MoveInstrumentComponent Child Algorithm");
+    g_log.error(e.what());
+  } catch (std::runtime_error &e) {
+    g_log.error(
+        "Unable to successfully run MoveInstrumentComponent Child Algorithm");
+    g_log.error(e.what());
+  }
 }
 
 /**
@@ -634,61 +685,49 @@ LoadHFIRSANS::getInstrumentDoubleParameter(const std::string &parameter) {
   std::vector<double> pars =
       m_workspace->getInstrument()->getNumberParameter(parameter);
   if (pars.empty()) {
-    g_log.warning() << "Parameter not found in the instrument parameter file: "
-                    << parameter << "\n";
+    g_log.warning() << "Parameter not found: " << parameter
+                    << " in the instrument parameter file.\n";
     return std::numeric_limits<double>::quiet_NaN();
   } else {
-    g_log.debug() << "Found the parameter in the instrument parameter file: "
-                  << parameter << " = " << pars[0] << "\n";
+    g_log.debug() << "Found the parameter: " << parameter << " = " << pars[0]
+                  << " in the instrument parameter file.\n";
     return pars[0];
   }
 }
-/**
- *  Source to Detector Distance is already calculated in the
-    metadata tag source_distance (if source_distance >= 0).
-    In the metadata we have:
-        source_distance
-        sample_aperture_to_flange
-        nguides
-    The nguides is the index to the Mantid table of number of guides <-> source
-    distances.
-    source_distance = MantidTable[nguides] - sample_aperture_to_flange.
- **/
+
 double LoadHFIRSANS::getSourceToSampleDistance() {
-  // First let's try to get source_distance first:
-  double sourceToSampleDistance =
-      boost::lexical_cast<double>(m_metadata["Header/source_distance"]);
-  // XML 1.03: source distance is now in meters
-  if (m_sansSpiceXmlFormatVersion >= 1.03) {
-    sourceToSampleDistance *= 1000; // convert to mm
+  const int nguides = static_cast<int>(
+      boost::lexical_cast<double>(m_metadata["Motor_Positions/nguides"]));
+  // m_workspace->run().getPropertyValueAsType<int>("number-of-guides");
+
+  // aperture-distances: array from the instrument parameters
+  std::string pars = getInstrumentStringParameter("aperture-distances");
+
+  double SSD = 0;
+  Mantid::Kernel::StringTokenizer tok(
+      pars, ",", Mantid::Kernel::StringTokenizer::TOK_IGNORE_EMPTY);
+  if (tok.count() > 0 && tok.count() < 10 && nguides >= 0 && nguides < 9) {
+    const std::string distance_as_string = tok[8 - nguides];
+    try {
+      auto distance_as_string_copy =
+          boost::algorithm::trim_copy(distance_as_string);
+      SSD = boost::lexical_cast<double>(distance_as_string_copy);
+    } catch (boost::bad_lexical_cast const &e) {
+      throw Kernel::Exception::InstrumentDefinitionError(
+          "Bad value for source-to-sample distance");
+    }
+  } else
+    throw Kernel::Exception::InstrumentDefinitionError(
+        "Unable to get source-to-sample distance");
+
+  // Check for an offset
+  double sourceSampleDistanceOffset =
+      getInstrumentDoubleParameter("source-distance-offset");
+  if (!std::isnan(sourceSampleDistanceOffset)) {
+    SSD += sourceSampleDistanceOffset;
   }
-  if (sourceToSampleDistance <= 0) {
-    g_log.warning()
-        << "Source To Sample Distance: Header/source_distance = "
-        << sourceToSampleDistance
-        << ". Trying to calculate it from the number of guides used and offset."
-        << '\n';
-    const int nGuides = static_cast<int>(
-        boost::lexical_cast<double>(m_metadata["Motor_Positions/nguides"]));
-    // aperture-distances: array from the instrument parameters
-    std::string guidesDistances =
-        getInstrumentStringParameter("aperture-distances");
-    std::vector<std::string> guidesDistancesSplit;
-    boost::split(guidesDistancesSplit, guidesDistances,
-                 boost::is_any_of("\t ,"), boost::token_compress_on);
-    sourceToSampleDistance =
-        boost::lexical_cast<double>(guidesDistancesSplit[nGuides]);
-    g_log.debug() << "Number of guides used = " << nGuides
-                  << " --> Raw SSD = " << sourceToSampleDistance << "mm.\n";
-    double sourceToSampleDistanceOffset = boost::lexical_cast<double>(
-        m_metadata["Header/sample_aperture_to_flange"]);
-    g_log.debug() << "SSD offset  = " << sourceToSampleDistanceOffset
-                  << "mm.\n";
-    sourceToSampleDistance -= sourceToSampleDistanceOffset;
-  }
-  g_log.information() << "Source To Sample Distance = "
-                      << sourceToSampleDistance << "mm.\n";
-  return sourceToSampleDistance;
+  g_log.debug() << "Source Sample Distance = " << SSD << ".\n";
+  return SSD;
 }
 
 /**
@@ -696,24 +735,40 @@ double LoadHFIRSANS::getSourceToSampleDistance() {
  * */
 void LoadHFIRSANS::setBeamDiameter() {
 
-  double sourceToSampleDistance = getSourceToSampleDistance();
-  addRunProperty<double>("source-sample-distance", sourceToSampleDistance,
-                         "mm");
+  double SourceToSampleDistance = 0.0;
+
+  try {
+    SourceToSampleDistance = getSourceToSampleDistance();
+    m_workspace->mutableRun().addProperty("source-sample-distance",
+                                          SourceToSampleDistance, "mm", true);
+    g_log.information() << "Computed SSD from number of guides: "
+                        << SourceToSampleDistance << " mm \n";
+  } catch (...) {
+    Mantid::Kernel::Property *prop =
+        m_workspace->run().getProperty("source-sample-distance");
+    Mantid::Kernel::PropertyWithValue<double> *dp =
+        dynamic_cast<Mantid::Kernel::PropertyWithValue<double> *>(prop);
+    SourceToSampleDistance = *dp;
+    g_log.warning() << "Could not compute SSD from number of guides, taking: "
+                    << SourceToSampleDistance << " mm \n";
+  }
+
   const double sampleAperture =
       boost::lexical_cast<double>(m_metadata["Header/sample_aperture_size"]);
   const double sourceAperture =
       boost::lexical_cast<double>(m_metadata["Header/source_aperture_size"]);
   g_log.debug() << "Computing beam diameter. m_sampleDetectorDistance="
                 << m_sampleDetectorDistance
-                << " SourceToSampleDistance=" << sourceToSampleDistance
+                << " SourceToSampleDistance=" << SourceToSampleDistance
                 << " sourceAperture= " << sourceAperture
                 << " sampleAperture=" << sampleAperture << "\n";
 
   const double beamDiameter = m_sampleDetectorDistance /
-                                  sourceToSampleDistance *
+                                  SourceToSampleDistance *
                                   (sourceAperture + sampleAperture) +
                               sampleAperture;
-  addRunProperty<double>("beam-diameter", beamDiameter, "mm");
+  m_workspace->mutableRun().addProperty("beam-diameter", beamDiameter, "mm",
+                                        true);
 }
 
 } // namespace DataHandling
